+++ ui/app/api/peers/route.ts
@@ -1,3 +1,4 @@
+import { getTruePeer } from '@/app/api/peers/getTruePeer';
 import {
   CatalogPeer,
   PeerConfig,
@@ -8,13 +9,10 @@ import prisma from '@/app/utils/prisma';
 import {
   BigqueryConfig,
   DBType,
-  EventHubConfig,
-  EventHubGroupConfig,
   Peer,
   PostgresConfig,
   S3Config,
   SnowflakeConfig,
-  SqlServerConfig,
 } from '@/grpc_generated/peers';
 import {
   CreatePeerRequest,
@@ -62,6 +60,8 @@ const constructPeer = (
       return;
   }
 };
+
+export const dynamic = 'force-dynamic';
 
 export async function POST(request: Request) {
   const body = await request.json();
@@ -116,55 +116,6 @@ export async function POST(request: Request) {
     }
   }
 }
-
-export const getTruePeer = (peer: CatalogPeer) => {
-  const newPeer: Peer = {
-    name: peer.name,
-    type: peer.type,
-  };
-  const options = peer.options;
-  let config:
-    | BigqueryConfig
-    | SnowflakeConfig
-    | PostgresConfig
-    | EventHubConfig
-    | S3Config
-    | SqlServerConfig
-    | EventHubGroupConfig;
-  switch (peer.type) {
-    case 0:
-      config = BigqueryConfig.decode(options);
-      newPeer.bigqueryConfig = config;
-      break;
-    case 1:
-      config = SnowflakeConfig.decode(options);
-      newPeer.snowflakeConfig = config;
-      break;
-    case 3:
-      config = PostgresConfig.decode(options);
-      newPeer.postgresConfig = config;
-      break;
-    case 4:
-      config = EventHubConfig.decode(options);
-      newPeer.eventhubConfig = config;
-      break;
-    case 5:
-      config = S3Config.decode(options);
-      newPeer.s3Config = config;
-      break;
-    case 6:
-      config = SqlServerConfig.decode(options);
-      newPeer.sqlserverConfig = config;
-      break;
-    case 7:
-      config = EventHubGroupConfig.decode(options);
-      newPeer.eventhubGroupConfig = config;
-      break;
-    default:
-      return newPeer;
-  }
-  return newPeer;
-};
 
 // GET all the peers from the database
 export async function GET(request: Request) {
